pyproject.toml
[project]
name = "mlb-odds-model"
version = "0.1.0"
description = "Modular MLB betting model focused on CLV and edge detection against DraftKings."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "pydantic-settings>=2.6.1",
  "pybaseball>=2.2.7",
  "python-dateutil>=2.9.0",
  "python-multipart>=0.0.12",
  "sqlmodel>=0.0.22",
  "tenacity>=9.0.0",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.7.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
